requirements.txt
flask-mongoengine
aniso8601==8.0.0 
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pylint-flask==0.6
pylint==2.5.2
pylint_flask_sqlalchemy==0.2.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
dnspython
Flask-PyMongo
flask_cors
scrapy
selenium
nltk
pythainlp
wheel
stop_words
sentence-transformers==1.2.0
annoy==1.17.0
cython==0.29.21
fuzzywuzzy==0.18.0
hdbscan==0.8.26
joblib==1.0.0
kiwisolver==1.3.1
# llvmlite==0.35.0
matplotlib==3.3.2
numpy==1.20.0
numba==0.52.0rc2
pandas==1.1.2
pillow==8.1.2
pyarrow
python-levenshtein==0.12.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.50.0
umap-learn==0.5.0
umap-learn[plot]
auxlib
conda
opencv-python==4.5.4.60
easyocr
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.1+cu113 
torchvision==0.11.2+cu113 
torchaudio===0.10.1+cu113 